highlight table rows
test if temp folder exists

+++ src/common/FileManager.java
@@ -9,6 +9,7 @@ public class FileManager {
     public static File temp = new File("temp");
 
     public static void removeTempFiles() throws IOException {
-        FileUtils.cleanDirectory(temp);
+        if (temp.exists())
+            FileUtils.cleanDirectory(temp);
     }
 }
